# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alphora API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/research-runs": {
      "post": {
        "tags": [
          "research-runs"
        ],
        "summary": "Create Research Runs",
        "operationId": "create_research_runs_api_research_runs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResearchRunsRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResearchRunSummary"
                  },
                  "title": "Response Create Research Runs Api Research Runs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "research-runs"
        ],
        "summary": "List Research Runs",
        "operationId": "list_research_runs_api_research_runs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ticker"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ResearchRunSummary"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/GroupedRuns"
                    }
                  ],
                  "title": "Response List Research Runs Api Research Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research-runs/cost-estimate": {
      "get": {
        "tags": [
          "research-runs"
        ],
        "summary": "Get Research Run Cost Estimate",
        "operationId": "get_research_run_cost_estimate_api_research_runs_cost_estimate_get",
        "parameters": [
          {
            "name": "strategy",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/StrategyEnum",
              "default": "funnel_research"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCostEstimate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research-runs/{run_id}": {
      "get": {
        "tags": [
          "research-runs"
        ],
        "summary": "Get Research Run",
        "operationId": "get_research_run_api_research_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research-runs/{run_id}/events": {
      "get": {
        "tags": [
          "research-runs"
        ],
        "summary": "Stream Research Run Events",
        "operationId": "stream_research_run_events_api_research_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research-runs/{run_id}/llm-calls": {
      "get": {
        "tags": [
          "research-runs"
        ],
        "summary": "List Research Run Llm Calls",
        "operationId": "list_research_run_llm_calls_api_research_runs__run_id__llm_calls_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LlmCallLogPublic"
                  },
                  "title": "Response List Research Run Llm Calls Api Research Runs  Run Id  Llm Calls Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research-runs/{run_id}/llm-calls/{log_id}/replay": {
      "post": {
        "tags": [
          "research-runs"
        ],
        "summary": "Replay Research Run Llm Call",
        "operationId": "replay_research_run_llm_call_api_research_runs__run_id__llm_calls__log_id__replay_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Log Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmCallReplayPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research-runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "research-runs"
        ],
        "summary": "Cancel Research Run",
        "operationId": "cancel_research_run_api_research_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/research-runs/{run_id}/resume": {
      "post": {
        "tags": [
          "research-runs"
        ],
        "summary": "Resume Research Run",
        "operationId": "resume_research_run_api_research_runs__run_id__resume_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/screeners/run": {
      "post": {
        "tags": [
          "screeners"
        ],
        "summary": "Run Screener",
        "operationId": "run_screener_api_screeners_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenerRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/screeners/runs/{screener_run_id}": {
      "get": {
        "tags": [
          "screeners"
        ],
        "summary": "Get Screener Run",
        "operationId": "get_screener_run_api_screeners_runs__screener_run_id__get",
        "parameters": [
          {
            "name": "screener_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Screener Run Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenerRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateResearchRunsRequest": {
        "properties": {
          "strategy": {
            "$ref": "#/components/schemas/StrategyEnum",
            "default": "tradingagents"
          },
          "trade_date": {
            "type": "string",
            "format": "date",
            "title": "Trade Date"
          },
          "tickers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 25,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Tickers"
          },
          "scope_payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MacroBriefScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "analysts": {
            "items": {
              "$ref": "#/components/schemas/AnalystKindEnum"
            },
            "type": "array",
            "title": "Analysts"
          },
          "llm_provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LlmProviderEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "debate_depth": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Debate Depth",
            "default": 3
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trade_date"
        ],
        "title": "CreateResearchRunsRequest"
      },
...
      "GroupedRuns": {
        "properties": {
          "queued": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunSummary"
            },
            "type": "array",
            "title": "Queued"
          },
          "running": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunSummary"
            },
            "type": "array",
            "title": "Running"
          },
          "recent": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunSummary"
            },
            "type": "array",
            "title": "Recent"
          },
          "failed": {
            "items": {
              "$ref": "#/components/schemas/ResearchRunSummary"
            },
            "type": "array",
            "title": "Failed"
          }
        },
        "type": "object",
        "required": [
          "queued",
          "running",
          "recent",
          "failed"
        ],
        "title": "GroupedRuns"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResearchRunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticker"
          },
          "strategy": {
            "$ref": "#/components/schemas/StrategyEnum"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatusEnum"
          },
          "final_rating": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinalRatingEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          },
          "scope_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Payload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ticker",
          "strategy",
          "status",
          "final_rating",
          "created_at"
        ],
        "title": "ResearchRunSummary"
      },
...
    }
  }
}